app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
blinker==1.7.0
    # via flask
click==8.1.7
    # via flask
flask==3.0.2
    # via
    #   -r requirements.in
    #   flask-swagger-ui
flask-swagger-ui==4.11.1
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
mysql-connector-python==8.3.0
    # via -r requirements.in
pyment==0.3.3
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
werkzeug==3.0.1
    # via flask
